+++ apps/wallet/src/ui/app/components/menu/content/index.tsx
@@ -11,13 +11,13 @@ import {
     useNavigate,
 } from 'react-router-dom';
 
+import { ImportLedgerAccounts } from '../../ledger/ImportLedgerAccounts';
 import { AccountsSettings } from './AccountsSettings';
 import { AutoLockSettings } from './AutoLockSettings';
 import { ExportAccount } from './ExportAccount';
 import { ImportPrivateKey } from './ImportPrivateKey';
 import MenuList from './MenuList';
 import { NetworkSettings } from './NetworkSettings';
-import { ImportLedgerAccounts } from './ledger/ImportLedgerAccounts';
 import { ErrorBoundary } from '_components/error-boundary';
 import {
     MainLocationContext,
